Add 0-print_list.c file

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,7 +6,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int list_len = 0;
+	size_t list_len = 0;
 
 	while (h)
 	{
